pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "share-the-wealth"
version = "0.1.0"
description = "Track politician trades and mirror them with your funds"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "yfinance>=0.2.36",
    "yahooquery>=2.3.0",
    "alpaca-py>=0.18.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "anthropic>=0.18.0",
]

[project.scripts]
stw = "share_the_wealth.cli:main"

[tool.setuptools.packages.find]
include = ["share_the_wealth*"]
